fix: booking & feat: new routes booking

+++ backend/src/controller/booking.js
@@ -1,9 +1,12 @@
 import { Booking } from "../model/booking.js";
+import { NotFound } from "../middleware/errors.js";
+import { Car } from "../model/car.js";
+import Garages from "../model/garage.js";
 
 export const getAllBookings = async (req, res, next) => {
   try {
     const booking = await Booking.findAll();
-    res.json(booking);
+    res.status(200).json({ booking: booking });
   } catch (error) {
     next(error);
   }
@@ -16,7 +19,36 @@ export const getBooking = async (req, res, next) => {
         id,
       },
     });
-    res.json(booking);
+    if (!booking) throw new NotFound("Booking not found");
+    res.status(200).json({ booking: booking });
+  } catch (error) {
+    next(error);
+  }
+};
+
+export const getBookingByCar = async (req, res, next) => {
+  try {
+    const { id } = req.params;
+    const car = await Car.findByPk(id);
+    if (!car) throw new NotFound("Car not found");
+
+    const booking = await Booking.findAll({ where: { id_car: id } });
+
+    res.status(200).json({ booking: booking });
+  } catch (error) {
+    next(error);
+  }
+};
+
+export const getBookingByGarage = async (req, res, next) => {
+  try {
+    const { id } = req.params;
+    const garage = await Garages.findByPk(id);
+    if (!garage) throw new NotFound("Garage not found");
+
+    const booking = await Booking.findAll({ where: { id_garage: id } });
+
+    res.status(200).json({ booking: booking });
   } catch (error) {
     next(error);
   }
@@ -31,7 +63,7 @@ export const createBooking = async (req, res, next) => {
       date_start: dateStart,
       date_end: dateEnd,
     });
-    res.json(newBooking);
+    res.status(201).json({ message: "Booking created", booking: newBooking });
   } catch (error) {
     next(error);
   }
@@ -42,12 +74,14 @@ export const updateBooking = async (req, res, next) => {
     const booking = await Booking.findOne({
       where: { id },
     });
-
+    if (!booking) throw new NotFound("Booking not found");
     booking.set(req.body);
 
     await booking.save();
 
-    res.json(booking);
+    return res
+      .status(200)
+      .json({ message: "Booking updated", booking: booking });
   } catch (error) {
     next(error);
   }
@@ -55,10 +89,17 @@ export const updateBooking = async (req, res, next) => {
 export const deleteBooking = async (req, res, next) => {
   const { id } = req.params;
   try {
+    const booking = await Booking.findOne({
+      where: {
+        id,
+      },
+    });
+    if (!booking) throw new NotFound("Booking not found");
+
     await Booking.destroy({
       where: { id },
     });
-    return res.sendStatus(204);
+    return res.status(200).send({ message: "Booking deleted" });
   } catch (error) {
     next(error);
   }
